Added support for a reverse proxy and reduced complexity of port mapping
This commit does two things: it removes the need for ports 4002 and 4003 by
listening for websocket connections on the same port of the mldonkey-next
webserver and ensures mldonkey-next webapp uses a proper url for the two
websocket channels.

Main mldonkey-next websocket is now under <mldonkey-next-domain>/ws.
Log stream websocket is now under <mldonkey-next-domain>/logstream.

This configuration simplifies the use with a reverse proxy from a different
network.